Use longs for QuickSort

+++ fjava/src/main/java/com/ikaver/aagarwal/common/problems/QuickSort.java
@@ -1,5 +1,5 @@
 package com.ikaver.aagarwal.common.problems;
 
 public interface QuickSort {
-  void sort(double [] array, int left, int right);
+  void sort(long [] array, int left, int right);
 }
